[Excel] 使用基于 MiniExcel 实现的 reader 导出数据, 耗时过长

+++ Excel/Collapsenav.Net.Tool.Excel.Core/Interface/IExcelContainer.cs
@@ -3,11 +3,11 @@ namespace Collapsenav.Net.Tool.Excel;
 public interface IExcelContainer<T> : IDisposable, IEnumerable<IEnumerable<T>>
 {
     IEnumerable<T> this[string field] { get; }
-    IEnumerable<T> this[long row] { get; }
-    T this[long row, long col] { get; }
-    T this[string field, long row] { get; }
+    IEnumerable<T> this[int row] { get; }
+    T this[int row, int col] { get; }
+    T this[string field, int row] { get; }
     IEnumerable<string> Headers { get; }
     IDictionary<string, int> HeadersWithIndex { get; }
-    long RowCount { get; }
+    int RowCount { get; }
     int Zero { get; }
 }
